build refactor
